+++ nachos/analyze.py
@@ -124,6 +124,18 @@ def get_arguments_parser():
     arguments_parser.add_argument(
         '-p', '--properties', help='which properties to carry out', required=True)
 
+    arguments_parser.add_argument(
+        '-I',
+        '--inverse',
+        action='store_true',
+        help='for vibrational contribution, show `property(total - current)` rather than `property(current)`')
+
+    arguments_parser.add_argument(
+        '-g',
+        '--group-vibs',
+        action='store_true',
+        help='Group vibrational contribution per perturbation order')
+
     return arguments_parser
 
 
@@ -176,7 +188,8 @@ def main():
 
     analyzer = analyzing.Analyzer(df, vibrational_contributions)
 
-    analyzer.analyze(properties, only=only, frequencies_to_show=frequencies)
+    analyzer.analyze(
+        properties, only=only, frequencies_to_show=frequencies, inverse_vibs=args.inverse, group_vibs=args.group_vibs)
 
 
 if __name__ == '__main__':