+++ tests/tests_nachos_bake.py
@@ -227,7 +227,7 @@ class BakeTestCase(NachosTestCase):
 
         stdout_t, stderr_t = process.communicate()
 
-        # self.assertEqual(len(stderr_t), 0, msg=stderr_t.decode())
+        self.assertEqual(len(stderr_t), 0, msg=stderr_t.decode())
         self.assertEqual(len(stdout_t), 0, msg=stdout_t.decode())
 
         self.assertTrue(os.path.exists(h5_path))
@@ -249,7 +249,7 @@ class BakeTestCase(NachosTestCase):
 
         stdout_t, stderr_t = process.communicate()
 
-        # self.assertEqual(len(stderr_t), 0, msg=stderr_t.decode())
+        self.assertEqual(len(stderr_t), 0, msg=stderr_t.decode())
         self.assertEqual(len(stdout_t), 0, msg=stdout_t.decode())
 
         self.assertTrue(os.path.exists(h5_path))
@@ -273,7 +273,7 @@ class BakeTestCase(NachosTestCase):
 
         stdout_t, stderr_t = process.communicate()
 
-        # self.assertEqual(len(stderr_t), 0, msg=stderr_t.decode())
+        self.assertEqual(len(stderr_t), 0, msg=stderr_t.decode())
         self.assertEqual(len(stdout_t), 0, msg=stdout_t.decode())
 
         self.assertTrue(os.path.exists(h5_path))